refactor che and tekton packages
