+++ pkg/controller/tektoninstallation/tektoninstallation_controller_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/codeready-toolchain/toolchain-operator/pkg/apis"
 	"github.com/codeready-toolchain/toolchain-operator/pkg/apis/toolchain/v1alpha1"
-	"github.com/codeready-toolchain/toolchain-operator/pkg/resources/tekton"
+	"github.com/codeready-toolchain/toolchain-operator/pkg/tekton"
 	"github.com/codeready-toolchain/toolchain-operator/pkg/test"
 	. "github.com/codeready-toolchain/toolchain-operator/pkg/test/olm"
 	. "github.com/codeready-toolchain/toolchain-operator/pkg/test/toolchain"
@@ -174,5 +174,5 @@ func apiScheme(t *testing.T) *runtime.Scheme {
 
 func newReconcileRequest(tektonInstallation *v1alpha1.TektonInstallation) reconcile.Request {
 	namespacedName := types.NamespacedName{Namespace: tektonInstallation.Namespace, Name: tektonInstallation.Name}
-	return reconcile.Request{namespacedName}
+	return reconcile.Request{NamespacedName: namespacedName}
 }
